requirements.txt
django==3.0.4
djangorestframework==3.11.0
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1